chore: 更新单测引入方式+picker提取pickerProps
