+++ src/packages/picker/index.taro.ts
@@ -1,5 +1,4 @@
 import Picker from './picker.taro'
 
-export type { PickerOption } from './types'
-export type { PickerProps } from './picker.taro'
+export type { PickerOption, PickerProps } from './types'
 export default Picker
