+++ src/packages/picker/index.ts
@@ -1,5 +1,4 @@
 import Picker from './picker'
 
-export type { PickerOption } from './types'
-export type { PickerProps } from './picker'
+export type { PickerOption, PickerProps } from './types'
 export default Picker
